requirements.txt
transformers
openai
langchain
PyYAML
python-dotenv
chromadb
sentence-transformers
-e .